Move current 2PC impl to v1 module
Moves the current implementation of the two-phase commit consensus
engine to a `v1` submodule. This will allow the addition of newer
versions that are not compatible with v1 while still keeping v1
available for backwards compatibility.
